Add git operations via HTTP
I'm leveraging Grack for the main functionality of git operations. Most
of the complexity here comes from user permissions.

The security here is not quite as good as I'd like, but as long as I'm
using Grack, I'm limited in the customizations I'm able to do.

For example, Grack raises a 403 if a user isn't allowed to access a
repository, and a 404 if the project doesn't exist. I'd like to raise
404 in both cases (although, 403 makes sense if the user only has read
access but is currently trying to write).

I'll try working on this in the future, but for now I believe this works
well enough.