Fix routes
Previously, navigating to `/this/path/does/not/exist` caused my
GrackWrapper to request a 401 and authenticate.

This only falls back to grack if the requested project name ends with
`.git`.

I also disabled any non-HTML requests for now. I don't plan on needing
it anytime soon, so I thought it would ease development for now. I'll
look into re-enabling it or finding alternatives when I need to.