Test both pushing and pulling permissions
This replaces the original `git clone` tests with tests that check for
both reads and writes.

I admit though that these tests aren't fully thorough. For example, it's
technically possible to give permissions to a user so that the user can
only push but not pull. I haven't tested those cases here though.